test(suite): move device tests to suite-common
