+++ suite-common/suite-utils/src/__fixtures__/device.ts
@@ -1,8 +1,7 @@
+import type { TrezorDevice } from '@suite-common/suite-types';
 import { testMocks } from '@suite-common/test-utils';
 import { DeviceModelInternal } from '@trezor/device-utils';
 import * as URLS from '@trezor/urls';
-
-import type { TrezorDevice } from 'src/types/suite';
 
 const { getSuiteDevice } = testMocks;
